Continuous page numbering on multiple source added.

+++ sources/Html2PdfServer/Renderer.h
@@ -21,7 +21,7 @@ typedef void (^RendererCompletionBlock)(NSData *, NSMutableDictionary *, NSArray
 
 - (id)initWithPool:(RendererPool *)pool;
 
-- (void)processRequest:(NSString *)requestUrl firstPageNumber:(int)firstPageNumber completionBlock:(RendererCompletionBlock)aBlock;
+- (void)processRequest:(NSString *)requestUrl firstPageNumber:(NSUInteger)firstPageNumber completionBlock:(RendererCompletionBlock)aBlock;
 
 - (void)startTask;
